Trim trailing zeros for DEC format
Force string conversion to number by multiplying by 1. This effectively
trims the trailing zeros of a decimal fraction.

This was already happening in the cases where we used Math.abs() - for
the same reason. Whether this is desirable behavior or not... you tell
me :-)

We can always add some explicit option for that, but for now I prefer no
trailing zeros.